requirements.txt
docker==7.1.0
nanodjango==0.9.2
openai==1.61.1
requests==2.32.3
svix-ksuid==0.6.2
tiktoken==0.8.0
wandb==0.19.6
weave==0.51.33